refactor: 장기 게임 패키지 분리
